requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
cycler==0.10.0
discord.py==1.6.0
idna==3.1
idna-ssl==1.1.0
kiwisolver==1.3.1
matplotlib==3.3.4
multidict==5.1.0
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.0
pyparsing==2.4.7
pytesseract==0.3.7
python-dateutil==2.8.1
six==1.15.0
twitchio==1.2.1
typing-extensions==3.7.4.3
websockets==8.1
yarl==1.6.3
